feat: Add convertInputData4Prisma utility function
This commit adds a new utility function `convertInputData4Prisma` to the `gqlQuery.ts` file. The function takes an input data object and converts it into the format expected by Prisma for updating records. It iterates over the keys of the input data object and creates a new object with each key mapped to `{ set: data[key] }`. This utility function enhances the flexibility of updating records in Prisma by providing a convenient way to convert input data into the required format.