feat: Add link to video edit page in user page
This commit adds a link to the video edit page in the user page. Now, when a user clicks on a video title, they will be redirected to the video edit page for that specific video. This enhances the user experience by providing a convenient way to access and edit videos directly from the user page.