feat: Add VideoCardList component to display a list of video cards
This commit adds the VideoCardList component, which is responsible for rendering a list of video cards. The component takes an array of video objects as props and maps over them to render individual VideoCard components. This new component improves code organization and reusability by separating the logic for rendering video cards into a separate component.