feat: Update UserAndVideosQuery to include Owner information
This commit modifies the UserAndVideosQuery in the generated types file to include the Owner information for each video. The Owner object now contains the id and name of the video owner. This change enhances the functionality of the query by providing access to the owner information along with the video metadata.